% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_spectrum.r
\name{average_spectrum}
\alias{average_spectrum}
\title{MS1 or MS2 spectra alignment}
\usage{
average_spectrum(splist, mz_window = 0.01)
}
\description{
Function used by process_consensus and process_lib2matrix for m/z alignment
}
