% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_spectrum.r
\name{cut_mz_list}
\alias{cut_mz_list}
\title{segment mz list}
\usage{
cut_mz_list(mzlist, mz_window)
}
\arguments{
\item{mzlist}{mass list}

\item{mz_window}{cut off window width}
}
\description{
segment mz list
}
