% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_SmartMS2.R
\name{denoise_ms2_spectrum}
\alias{denoise_ms2_spectrum}
\title{Keep top peaks}
\usage{
denoise_ms2_spectrum(sp, mz0, max_peak, min_relative, normalized = T)
}
\arguments{
\item{sp}{spectra}

\item{mz0}{pepmass}

\item{max_peak}{maxial number o peaks to be retained}

\item{min_relative}{minimal relative intensity to be retained}

\item{normalized}{boolean}
}
\description{
Keep top peaks
}
