% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_consensus.R
\name{denoise_spectrum}
\alias{denoise_spectrum}
\title{Keep top peaks and noramalzie}
\usage{
denoise_spectrum(sp, max_peak, min_relative)
}
\arguments{
\item{sp}{spectrum}
}
\description{
Keep top peaks and noramalzie
}
