% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{library2}
\alias{library2}
\title{A test data contains spectra and metadata
 @docType data}
\format{
An object of class \code{list} of length 2.
}
\usage{
data("DRUG_THERMO_LIBRARY")
}
\description{
A test data contains spectra and metadata
 @docType data
}
\examples{
data("DRUG_THERMO_LIBRARY")
library2

}
\keyword{datasets}
