% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library2matrix.R
\name{library2matrix}
\alias{library2matrix}
\title{Generate aligned fragment and neutral loss matrix from MS2 spectral library}
\usage{
library2matrix(input_library, consensus_window = 0.02)
}
\arguments{
\item{input_library}{A list generated by the function library_generator() or the name of mgf/msp/RData spectral library file}

\item{mz_window}{m/z tolerance window (in da) for spectra alignment}
}
\description{
The function reduces the spectral library to improve search efficiency
}
