% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library_combiner.R
\name{library_combiner}
\alias{library_combiner}
\title{Combine multiple libraries}
\usage{
library_combiner(...)
}
\arguments{
\item{library1}{First library. A list of object from library_generator, or library file names (mgf, rdata or msp format)}

\item{library2}{Second library}
}
\description{
The function generates a combined library from several individual libraries
}
\examples{

\donttest{data(DRUG_THERMO_LIBRARY)
combined_lib = library_combiner(library1, library2)}


}
