% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library_manager.R
\name{library_manager}
\alias{library_manager}
\title{Searching and managing the spectral library}
\usage{
library_manager(library, query = "", logical = c("AND", "OR"),
  ppm_search = 20, rt_search = 12)
}
\arguments{
\item{library}{A list generated by the function library_generator() or the name of mgf spectral library file}

\item{query}{Vector of characters. Vector of conditions used for querying the library. e.g. c("IONMODE=Positive","PEPMASS=325.19"). The left-hand side must match with the medata items of the searched library.}

\item{logical}{Character. "AND" for selecting scans that satisfy all conditions, "OR" when selecting records that satisfy at least one condition}

\item{ppm_search}{Numeric. Mass tolerance in ppm. Only used when searching by precursor mass "PEPMASS=..."}

\item{rt_search}{Numeric. Retention time tolerance in second (although rt in the query and metadata in min). Only used when searching by retention time "RT=..."}
}
\value{
\itemize{
 \item{SELECTED:}{ Library object that only contain selected scans}
 \item{ID_SELECTED:}{ IDs of selected scans}
 \item{LEFT:}{ Library object that only contain unselected scans}
 \item{ID_SELECTED:}{ IDs of unnselected scans}
}
}
\description{
The function queries and selects or removes scans from the spectral library that satisfy user-defined conditions (query metadata)
}
\examples{

data(DRUG_THERMO_LIBRARY)

# Search library using query command lines:
query = library_manager(library2,query=c("IONMODE=Positive","RT=1.2"), logical="AND", rt_search=6)

# Create a new library from query:
new_library1 = query$SELECTED

# Summary of found compounds:
library_reporter(new_library1)

# Remove scans from current library according to query:
new_library2 = query$LEFT

# Add another filter:
query = library_manager(new_library1,query=c("IONMODE=Positive","MSLEVEL=2","RT=1.2"))
new_library3 = query$SELECTED

}
