% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library_merger.R
\name{library_merger}
\alias{library_merger}
\title{Extracting or generating representative scans for each compound and create a new library}
\usage{
library_merger(
  input_library,
  method = c("most_recent", "max_tic", "consensus", "common_peaks"),
  consensus_window = 0.01,
  params.ms.preprocessing = list(relative = 0.01, max_peaks = 200)
)
}
\arguments{
\item{input_library}{A list generated by the function library_generator() or the name of mgf/msp/RData spectral library file}

\item{method}{A character of method for grouping library by compound IDs.
\itemize{
  \item{consensus:}{ Consensus spectra were generated by merging spectra of the same compound ID. All peaks were kept, similar peaks were aligned.}
  \item{common_peaks:}{ Consensus spectra were generated by merging spectra (MS1 and MS2) of the same compound ID. Peaks detected in all spectra were kept and averaged.}
  \item{max_tic:}{ For the same compound ID, the spectrum with higher TIC was extracted and kept.}
  \item{most_recent:}{ For the same compound ID, the most recent record was kept.}
}}

\item{consensus_window}{m/z window (in Dalton) for spectra alignment, only used when method = "consensus" or "common_peaks". To generate consensus spectra, mass peaks in different spectra within the mass window were aligned by averaging their mass values and intensities. The metadata was kept for the spectrum with higher TIC.}

\item{params.ms.preprocessing}{Paremters for post-processing merged scan. It must contain:
\itemize{
 \item{relative:}{ Numeric between 0 and 100. Relative intensity threshold of the highest peak in each spectrum.}
 \item{max_peaks:}{ Integer higher than 3. Maximum number of peaks kept per spectrum from the highest peak.}
}}
}
\value{
The same as library_generator()
}
\description{
The function generates a consensus scan for each compound ID in the spectral library.
}
\examples{

data(DRUG_THERMO_LIBRARY)

# Withholding the scan (MS1 and MS) with highest TIC for each ID:

library2_1 = library_merger(library2, method="max_tic")

# Generating consensus spectra for all scans (MS1 and MS) of the same ID. The scan with highest TIC is used for metadata:

library2_2 = library_merger(library2, method="common_peaks", consensus_window = 0.005)

}
\author{
Youzhong Liu, \email{YLiu186@ITS.JNJ.com}
}
