% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library_messar.R
\name{library_messar}
\alias{library_messar}
\title{Matching query spectrum to existing library}
\usage{
library_messar(library, ppm_search = 20, tops = 3, min.F1 = 0,
  output = "library_messar.mgf")
}
\arguments{
\item{library}{A list generated by the function library_generator() or the name of mgf spectral library file}

\item{ppm_search}{Numeric. Mass tolerance in ppm for precursor/fragment search.}

\item{tops}{Integer. Top recommended substructures in the output mgf file}

\item{output}{String. File name of output library.}
}
\value{
\itemize{
   \item{rules_extracted:}{ Matched rule list}
   \item{sub_extracted:}{ Library object that contain found scans.}
   \item{MGF FILE:}{ A MGF file with annotated substructures will be written in user's folder.}
}
}
\description{
The function searches unknown MS/MS spectra in a spectral library
}
\examples{

data(DRUG_THERMO_LIBRARY)
results = library_messar(library2, ppm_search = 20, tops = 5, output = "library2_messar.mgf")

}
\author{
Youzhong Liu, \email{Youzhong.Liu@uantwerpen.be}
}
