% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library_messar2.R
\name{library_messar2}
\alias{library_messar2}
\title{Matching query spectrum to existing library}
\usage{
library_messar2(library, type = c("drug", "metabolite"), tops = 5,
  max_peaks = 200, min_relative = 0.1,
  output = "library_messar_drug.mgf")
}
\arguments{
\item{library}{A list generated by the function library_generator() or the name of mgf spectral library file}

\item{tops}{Integer. Top recommended substructures in the output mgf file}

\item{max_peaks}{Integer. Top most intense peaks kept in the spectra}

\item{min_relative}{Numeric between 0.1 and 1. Minimum relative intensity of peaks to not be considered as noise.}

\item{output}{String. File name of output library.}
}
\value{
\itemize{
   \item{library_annotated:}{ Library annotated}
   \item{sub_extracted:}{ Library object that contain found scans.}
   \item{MGF FILE:}{ A MGF file with annotated substructures will be written in user's folder.}
}
}
\description{
The function searches unknown MS/MS spectra in a spectral library
}
\examples{

data(DRUG_THERMO_LIBRARY)
results = library_messar2(library2, type = "drug", tops = 5, output = "thermo_substructure.mgf")

}
\author{
Youzhong Liu, \email{Youzhong.Liu@uantwerpen.be}
}
