% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library_messar_drug.R
\name{library_messar_drug}
\alias{library_messar_drug}
\title{Matching query spectrum to existing library}
\usage{
library_messar_drug(library, ppm_search = 10, tops = 3,
  max_peak = 200, min_relative = 0.1,
  output = "library_messar_drug.mgf")
}
\arguments{
\item{library}{A list generated by the function library_generator() or the name of mgf spectral library file}

\item{ppm_search}{Numeric. Mass tolerance in ppm for precursor/fragment search.}

\item{tops}{Integer. Top recommended substructures in the output mgf file}

\item{output}{String. File name of output library.}
}
\value{
\itemize{
   \item{rules_extracted:}{ Matched rule list}
   \item{sub_extracted:}{ Library object that contain found scans.}
   \item{MGF FILE:}{ A MGF file with annotated substructures will be written in user's folder.}
}
}
\description{
The function searches unknown MS/MS spectra in a spectral library
}
\examples{

data(DRUG_THERMO_LIBRARY)
results = library_messar_drug(library2, ppm_search = 10, tops = 5, output = "library_messar_drug.mgf")

}
\author{
Youzhong Liu, \email{Youzhong.Liu@uantwerpen.be}
}
