% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library_messar_simple.R
\name{library_messar_simple}
\alias{library_messar_simple}
\title{Substructure recommendation for unknown spectrum}
\usage{
library_messar_simple(query_spectrum = NULL, prec_mz = 0,
  type = c("drug", "metabolite"), tops = 5, max_peaks = 200,
  min_relative = 0.1)
}
\arguments{
\item{query_spectrum}{Two-column data matrix. Two columns represent m/z and intensity of query tandem spectrum. At least 3 valid peaks should be provided}

\item{prec_mz}{Numeric. Precursor mass of query spectrum (strongly recommended). Please set to 0 if unknown.}

\item{type}{Character. Two algorithms focused either on "drug" or "metabolite"}

\item{tops}{Integer. Top recommended substructures}

\item{max_peaks}{Integer. Top most intense peaks kept in the spectra}

\item{min_relative}{Numeric between 0.1 and 1. Minimum relative intensity of peaks to not be considered as noise.}
}
\value{
Recommended substructures (ordered) as a two-column matrix: SMILES code and score
}
\description{
The function recommends substructures for an unknown spectrum
}
\examples{

url_test = "https://raw.githubusercontent.com/daniellyz/MergeION2/master/inst/caffeine.txt"

test = read.table(url_test, header = F, sep='\\t')

results = library_messar_simple(query_spectrum = test, prec_mz = 195.088, type = "drug", tops = 3)

}
\author{
Youzhong Liu, \email{Youzhong.Liu@uantwerpen.be}
}
