% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library_query.R
\name{library_query}
\alias{library_query}
\title{Spectral library query}
\usage{
library_query(
  input_library = NULL,
  query_expression = "Polarity = Positive",
  query_spectrum = NULL,
  params.search = list(mz_search = 0.01, ppm_search = 10, rt_seach = 15),
  params.query.sp = list(prec_mz = 100, use_prec = T, polarity = "Positive", method =
    "Cosine", min_frag_match = 6, reaction_type = "Metabolic")
)
}
\arguments{
\item{input_library}{Character or a list object. If character, file name with extension mgf, msp or RData. If list, must contain "complete", "consensus" and "network".}

\item{query_expression}{Vector of characters. Vector of conditions used for querying the library. e.g. c("IONMODE=Positive","PEPMASS=325.19"). The left-hand side must match with the metadata items of the searched library.}

\item{query_spectrum}{Two-column data matrix. Two columns represent m/z and intensity of query tandem spectrum. At least 3 valid peaks should be provided}

\item{params.search}{General parameters for searching spectral library
\itemize{
 \item{mz_search:}{ Numeric. Absolute mass tolerance in Da for fragment match.}
 \item{ppm_search:}{ Numeric. Mass tolerance in ppm for precursor mass match.} 
 \item{rt_search:}{ Numeric. Absolute retention time tolerance in second.}
}}

\item{params.query.sp}{Parameters for matching an unknown spectrum
\itemize{
  \item{prec_mz:}{ Numeric. Precursor mass of query spectrum.}
  \item{use_prec:}{ Boolean. If set to TRUE, precursor mass is used to "pre-query" the library.}
  \item{polarity:}{ Character. Ion mode of the query spectrum, either positive or negative.}
  \item{method:}{ Character. Similarity metrics.}
   \enumerate{
       \item{Matches:}{ Conting number of fragment or neutral loss matches}
       \item{Dot:}{ Dot similarity score based on intensity vectors of fragments.}
       \item{Cosine:}{ COsine similarity score based on intensity vectors of fragments.}
       \item{Spearman:}{ Spearman similarity based on intensity ranks of fragments.}
       \item{MassBank:}{Similarity score used by MassBank using weighted pearson score}
       \item{NIST:}{Similarity score used by NIST using weighted pearson score}}
   \item{min_frag_match}{ Minimum matched peaks (or corresponding neutral losses) to allow a match.}
   \item{reaction_type:}{ Character. Either "Metabolic" and "Chemical". Type of transformation list used to annotate mass difference between connected features in molecular network.}
}}
}
\value{

}
\description{
The function queries the input spectral library based on a query expression and/or a query MS/MS spectrum
}
\examples{

}
\author{
Youzhong Liu, \email{Youzhong.Liu@uantwerpen.be}
}
