% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library_reporter.R
\name{library_reporter}
\alias{library_reporter}
\title{Generating a summary of the spectral library}
\usage{
library_reporter(library)
}
\arguments{
\item{library}{A list generated by the function library_generator() or the name of mgf spectral library file}
}
\description{
The function generates a report from input spectral library
}
\examples{

data(DRUG_THERMO_LIBRARY)
library_reporter(library2)

}
