% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library_search_engine.R
\name{library_search_engine}
\alias{library_search_engine}
\title{Matching query spectrum to existing library}
\usage{
library_search_engine(
  library_type = c("Metabolite", "Private", "Public", "Local"),
  local_library = NULL,
  query_spectrum = NULL,
  ion.mode = c("Positive", "Negative"),
  method = c("Fragment", "Simple", "Cosine"),
  prec_mz = 0,
  use.prec = FALSE,
  ppm_search = 10,
  tops = 5,
  max_peaks = 200,
  min_relative = 0.1,
  mirror.plot = T,
  png.out = F
)
}
\arguments{
\item{library_type}{A character precising the type of library. Fast search: "Metabolite" for precompiled GNPS + MASSBANK library; "Private" for precompiled JANSSEN library (2019-9-20); "Public" for drug spectra in public domains (GNPS and MASSBANK, positive mode only); Slower search: "Local" if a customized spectra library is provided}

\item{local_library}{Customized spectral library. Obligatory if library_type = "Local". A list generated by the function library_generator() or the name of mgf spectral library file}

\item{query_spectrum}{Two-column data matrix. Two columns represent m/z and intensity of query tandem spectrum. At least 3 valid peaks should be provided}

\item{ion.mode}{Character indicating whether the query spectrum is in Positive or Negative mode}

\item{method}{Character. Method to compare the similarity between query spectrum and library spectra.
\itemize{
  \item{Fragment:}{ Counting only number of fragment matches}
  \item{Simple:}{ Conting number of fragment and neutral loss matches}
  \item{Cosine:}{ Cosine similarity score from OrgMassSpecR package.}
}}

\item{prec_mz}{Numeric. Precursor mass of query spectrum (if known). Default value is 0. Must NOT be 0 if method = "All" or use.prec = TRUE (see below)}

\item{use.prec}{Boolean. If set to TRUE, precursor mass (if known) is used to "pre-query" the library}

\item{ppm_search}{Numeric. Mass tolerance in ppm for precursor search.}

\item{tops}{Integer. Number of top compound candidates kept in the output mgf file}

\item{max_peaks}{Integer. Top most intense peaks kept in the query spectrum}

\item{min_relative}{Numeric between 0.1 and 1. Minimum relative intensity of peaks to not be considered as noise in the query spectrum}

\item{mirror.plot}{Boolean. True if the query-library comparison is visualized as mirror plot}

\item{png.out}{Boolean. True if plotted mirror spectra are exported as png images!}
}
\value{
\itemize{
   \item{<plot>:}{ Comparing query spectrum to ordered "hits" in the spectrum library}
   \item{SELECTED:}{ Library object that contain found scans.}
   \item{ID_SELECTED:}{ IDs of found compounds.}
   \item{SCORES:}{ Similarity scores between query spectrum and each }
}
}
\description{
The function searches unknown MS/MS spectra in a spectral library
}
\examples{

dat = cbind(c(136.073,149.071,151.0991,180.047),c(1,1,20,3))

# Local search:
data(DRUG_THERMO_LIBRARY) # Load library2
query = library_search_engine(library_type = "Local", local_library = library2, 
query_spectrum = dat, method = "Cosine", prec_mz = 0, use.prec =FALSE, mirror.plot = F)

# Search in public database:
query = library_search_engine(library_type = "Public", local_library = NULL, 
query_spectrum = dat, method = "Simple", prec_mz = 346.122, use.prec =FALSE, mirror.plot = F)

# Query-library comparison via mirror plot:
library_visualizer_similarity(query$SELECTED, id= query$ID_SELECTED[1], query_spectrum = dat)

}
\author{
Youzhong Liu, \email{Youzhong.Liu@uantwerpen.be}
}
