% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library_search_engine_batch.R
\name{library_search_engine_batch}
\alias{library_search_engine_batch}
\title{Matching query spectra to existing library}
\usage{
library_search_engine_batch(
  library_type = c("Metabolite", "Private", "Public", "Local"),
  local_library = NULL,
  query_library = NULL,
  method = c("Fragment", "Simple", "Cosine"),
  use.prec = FALSE,
  ppm_search = 10,
  tops = 5,
  max_peaks = 200,
  min_relative = 0.1,
  output = "annotation.mgf"
)
}
\arguments{
\item{library_type}{A character precising the type of library. Fast search: "Metabolite" for precompiled GNPS + MASSBANK library; "Private" for precompiled JANSSEN library (2019-9-20); "Public" for drug spectra in public domains (GNPS and MASSBANK, positive mode only); Slower search: "Local" if a customized spectra library is provided}

\item{local_library}{Customized spectral library. Obligatory if library_type = "Local". A list generated by the function library_generator() or the name of mgf spectral library file}

\item{query_library}{A list generated by the function library_generator() or the name of mgf spectral library file}

\item{method}{Character. Method to compare the similarity between query spectrum and library spectra.
\itemize{
  \item{Fragment:}{ Counting only number of fragment matches}
  \item{Simple:}{ Conting number of fragment and neutral loss matches}
  \item{Cosine:}{ Cosine similarity score from OrgMassSpecR package.}
}}

\item{use.prec}{Boolean. If set to TRUE, precursor mass (PEPMASS in the mgf file) is used to "pre-query" the library}

\item{ppm_search}{Numeric. Mass tolerance in ppm for precursor search.}

\item{tops}{Integer. Number of top compound candidates kept in the output mgf file}

\item{max_peaks}{Integer. Top most intense peaks kept in eacg query spectrum}

\item{min_relative}{Numeric between 0.1 and 1. Minimum relative intensity of peaks to not be considered as noise in the query spectrum}

\item{output}{String. File name of output library with annotation result.}
}
\value{
\itemize{
   \item{library_annotated:}{ Library annotated}
   \item{sub_extracted:}{ Library object that contain found scans.}
   \item{MGF FILE:}{ A MGF file with annotated substructures will be written in user's folder.}
}
}
\description{
The function batch searches unknown MS/MS spectra (written in a mgf file) in a spectral library
}
\examples{

data(DRUG_THERMO_LIBRARY)

# Search in public database:
query = library_search_engine_batch(library_type = "Public", local_library = NULL, 
query_library = library2, method = "Simple", use.prec =FALSE)

}
\author{
Youzhong Liu, \email{Youzhong.Liu@uantwerpen.be}
}
