% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_visualizer.R
\name{library_visualizer_interactive}
\alias{library_visualizer_interactive}
\title{Interactive spectral plot}
\usage{
library_visualizer_interactive(
  plotData,
  x = "m/z",
  y = "Intensity",
  colorVar = "ID",
  facetVar = "MSLEVEL",
  title = ""
)
}
\arguments{
\item{plotData}{A data frame used to plot}

\item{x}{x-axis variable quote or unquoted or column index}

\item{y}{y-axis variable or column index}

\item{facetVar}{the variable used to facet the graph by row or column index}

\item{coloVar}{color variable or column index}
}
\description{
Interactive spectral plot
}
\author{
Yingjie Zhang
}
