% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library_visualizer_similarity.R
\name{library_visualizer_similarity}
\alias{library_visualizer_similarity}
\title{Visualizing query spectrum matched to existing library}
\usage{
library_visualizer_similarity(library, id = library$metadata$ID[1],
  query_spectrum = NULL, png.out = F)
}
\arguments{
\item{library}{A list generated by the function library_generator() or library_similarity()$SELECTED}

\item{query_spectrum}{Two-column data matrix. Two columns represent m/z and intensity of query tandem spectrum}

\item{png.out}{Boolean. True if plotted mirror spectra are exported as png images!}
}
\value{
Comparing query spectrum to "hits" in the spectrum library via mirror plot
}
\description{
The function generates the comparison plot between query spectrum and library
}
\author{
Youzhong Liu, \email{Youzhong.Liu@uantwerpen.be}
}
