% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library_writer.R
\name{library_writer}
\alias{library_writer}
\title{Writing the spectral library to an open format}
\usage{
library_writer(output_library, con = "output_library.mgf")
}
\arguments{
\item{output_library}{List that contains at list 2 elements.
\itemize{
  \item{sp:}{ List of all extracted spectra. Each spectrum is a data matrix with two columns: m/z and intensity}
  \item{metadata:}{ Data frame containing metadata of extracted scans. PEPMASS and RT are updated based on actually-detected scans. Following five columns are added: FILENAME, MSLEVEL, TIC, MASS_DEV, SCANNUMBER and SCANS}
}}

\item{con}{Name of the output library, the file extension must be mgf, msp or rdata}
}
\description{
The function writes the output of library_generator into mgf, msp or rdata file.
}
\examples{

data(DRUG_THERMO_LIBRARY)
library_writer(library2,"library_V2_bis.mgf")

}
