% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_editor.R
\name{metadata_editor}
\alias{metadata_editor}
\title{Create additional metadata for library generation}
\usage{
metadata_editor(
  ref,
  processing.algorithm = c("Default", "compMS2Miner", "RMassBank"),
  polarity = c("Positive", "Negative"),
  add.adduct = T
)
}
\description{
Function used by library_generator() to generate input metadata
}
