% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_MS1_screener.R
\name{metadata_MS1_screener}
\alias{metadata_MS1_screener}
\title{Create pseudo-metadata for library generation}
\usage{
metadata_MS1_screener(raw_data_file, screener = c("XCMS", "DDA"),
  ref = NULL, max.charge = 1, ppm_search = 20, rt_search = 12,
  baseline = 1000, snthreshold = 30, min.points = 8)
}
\description{
Function used by library_generator to extract LC-MS features from MS1 chromatogram of DDA/targeted-mode
}
