% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_generator.R
\name{network_generator}
\alias{network_generator}
\title{Create feature-based molecular networks}
\usage{
network_generator(
  raw_data_file,
  feature_table = NULL,
  input_library = NULL,
  input_matrix = NULL,
  polarity = c("Positive", "Negative"),
  params.screening = list(baseline = 1000, relative = 0.01, max_peaks = 200),
  params.search = list(mz_search = 0.005, ppm_search = 10, rt_search = 15),
  params.similarity = list(method = "Pearson", min.frag.match = 6, min.score = 0.6,
    topK = 10),
  params.network = list(method = "Spearman", min.abs.cor = 0.4, use.reaction = TRUE)
)
}
\arguments{
\item{raw_data_file}{Character. The LC-MS/MS file in mzML/mzXML format or MS/MS spectral summary file in mgf format. Check following link for more information about MS/MS spectral summary file: https://ccms-ucsd.github.io/GNPSDocumentation/featurebasedmolecularnetworking-with-mzmine2/}

\item{feature_table}{Character. The feature quantification table file. The file should be tab, comma or semi-colon separated txt, dat or csv format. It should contain at least 3 columns: LC-MS feature ID, m/z and RT that could be detected in the raw data file. It can also contain abundance of these features in samples. Check following link for more details about the format: https://ccms-ucsd.github.io/GNPSDocumentation/featurebasedmolecularnetworking-with-mzmine2/}

\item{input_library}{A list generated by the function library_generator() or the name of mgf/msp/RData spectral library file.}

\item{input_matrix}{Transformed spectral library by the function library2matrix. Used for MassBank or GNPS spectra library search. Please set to NULL otherwise}

\item{polarity}{character. Either "Positive" or "Negative". Ion mode of the LC-MS/MS file.}

\item{params.screening}{Parameters for feature screening and MS2 spectra pre-processing from raw_data_file:
\itemize{
 \item{baseline:}{ Numeric. Absolute intensity threshold that is considered as a mass peak and written into the library.}
 \item{relative:}{ Numeric between 0 and 100. Relative intensity threshold of the highest peak in each spectrum, peaks above both absolute and relative thresholds are saved in the library.}
 \item{max_peaks:}{ Integer higher than 3. Maximum number of peaks kept per spectrum from the highest peak.}
}}

\item{params.search}{List of parameters for feature screening, combining precursor ions and fragments in the input file, as well as for searching in the spectral library. The list must contain following elements:
\itemize{
 \item{mz_search:}{ Numeric. Absolute mass tolerance in Da.}
 \item{ppm_search:}{ Numeric. Absolute mass tolerance in ppm.} 
 \item{rt_search:}{ Numeric. Absolute retention time tolerance in second.}
}}

\item{params.network}{Parameters for post-filtering and annotation of network edges: based on feature correlation (if feature quantification table is provided) and mass difference
\itemize{
 \item{method:}{ Characeter. Correlation metrics. Must be "Pearson" or "Spearman".}
 \item{min.abs.cor:}{Numeric between 0 and 1. Minimum absolute correlation for kept edges.}
 \item{use.reaction:}{ Boolean. TRUE if keep only edges whose mass difference can be annotated to known biochemical reactions.}
}}
}
\description{
Generate feature-based molecular networks from raw LC-MS/MS file or spectral summary file, automatically annotate features in a spectral library
}
