% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_compMS2Miner.R
\name{process_compMS2Miner}
\alias{process_compMS2Miner}
\title{Read and combine targeted MS1 and MS2 scans from one LC-MS/MS file using compMS2Miner algorithm}
\usage{
process_compMS2Miner(
  mzdatafiles = NULL,
  ref = NULL,
  polarity = c("Positive", "Negative"),
  include.MS1 = FALSE,
  rt_search = 10,
  rt_gap = 30,
  ppm_search = 10,
  mz_search = 0.01,
  baseline = 1000,
  relative = 5,
  max_peaks = 200,
  normalized = T
)
}
\description{
Function used by library_generator
}
