% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_consensus.R
\name{process_consensus}
\alias{process_consensus}
\title{Extracting or generating representative or consensus scans for each compound ID}
\usage{
process_consensus(
  input_library,
  method = c("most_recent", "consensus", "consensus2", "common_peaks"),
  consensus_window = 0.01,
  relative = 0.01,
  max_peaks = 200,
  IDsUpdated = NULL
)
}
\arguments{
\item{input_library}{Character or a list object. If character, name of the existing library into which new scans are added, the file extension must be mgf, msp or RData; please set to NULL if the new library has no dependency with previous ones.}

\item{IDsUpadated}{new compound ID which were updated in the \bold{complete} library yet to update consensus library for.}
}
\description{
Extracting or generating representative or consensus scans for each compound ID
}
