% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_dda.R
\name{process_dda}
\alias{process_dda}
\title{Create pseudo-library file from a LC-MS/MS data file through non-targeted feature screening}
\usage{
process_dda(
  raw_data_file,
  polarity = c("Positive", "Negative"),
  ppm_search = 10,
  rt_search = 12,
  baseline = 1000
)
}
\arguments{
\item{raw_data_file}{Character. The LC-MS/MS file name from which fragmented LC-MS features are extracted. All files must have be in centroid-mode with mzML or mzMXL extension!}

\item{polarity}{character. Either "Positive" or "Negative". Ion mode of the LC-MS/MS file.}

\item{ppm_search}{Numeric. Absolute mass tolerance in ppm for feature screening and extraction}

\item{rt_search}{Numeric. Absolute retention tolerance in seconds for feature screening and extraction}

\item{baseline}{Numeric. Absolute intensity threshold above which is considered as a mass peak, otherwise noise.}
}
\description{
Function used by library_generator to screen LC-MS features from unknown mzML/mzXML file
}
