% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_mzmine.R
\name{process_mzmine}
\alias{process_mzmine}
\title{Create metadata from mzmine feature quantification table}
\usage{
process_mzmine(
  mzmine_file,
  polarity = c("Positive", "Negative"),
  combine.adduct = FALSE,
  remove.halogene = FALSE,
  remove.minor.coelution = FALSE,
  mz_search = 0.01,
  rt_search = 0.2
)
}
\description{
Function used to build molecular networks
}
