% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_query.R
\name{process_query}
\alias{process_query}
\title{Searching and managing the spectral library}
\usage{
process_query(library0, query = "", ppm_search = 20, rt_search = 12)
}
\description{
The function is used by library_query() to select scans based on a query expression.
}
