% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_RMassBank.R
\name{process_RMassBank}
\alias{process_RMassBank}
\title{Read and combine targeted MS1 and MS2 scans from one LC-MS/MS file using RMassBank algorithm}
\usage{
process_RMassBank(
  mzdatafiles = NULL,
  ref = NULL,
  polarity = c("Positive", "Negative"),
  include.MS1 = FALSE,
  add.adduct = FALSE,
  rt_search = 10,
  ppm_search = 10,
  baseline = 1000,
  relative = 5,
  max_peaks = 200,
  recalibration = F,
  normalized = T
)
}
\description{
Function used by library_generator
}
