% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_similarity.R
\name{process_similarity}
\alias{process_similarity}
\title{Searching a query spectrum in a spectral library
The function calculates spectral similarity of a query spectrum to an existing spectral library}
\usage{
process_similarity(
  query_spectrum,
  polarity = "Positive",
  prec_mz = 100,
  use.loss = TRUE,
  use.prec = FALSE,
  input_library = NULL,
  method = c("Precision", "Recall", "F1", "Cosine", "Spearman", "MassBank", "NIST", "HM",
    "Entropy"),
  prec_ppm_search = 10,
  frag_mz_search = 0.005,
  min_frag_match = 6
)
}
\arguments{
\item{query_spectrum}{two columns spectrum list}

\item{use.loss}{boolean if natrual loss also be used, \code{prec_mz} can not be missing if set true.}

\item{input_library}{Character or a list object. If character, file name with extension mgf, msp or RData. If list, must contain "complete", "consensus" and "network".}
}
\description{
Searching a query spectrum in a spectral library
The function calculates spectral similarity of a query spectrum to an existing spectral library
}
\author{
Youzhong Liu, \email{Youzhong.Liu@uantwerpen.be}
}
