% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_SmartMS2.R
\name{process_SmartMS2}
\alias{process_SmartMS2}
\title{Read and combine targeted MS2 scans from one LC-MS/MS file using default MergeION algorithm}
\usage{
process_SmartMS2(
  mzdatafiles = NULL,
  ref = NULL,
  rt_search = 10,
  rt_gap = 30,
  ppm_search = 10,
  mz_search = 0.01,
  baseline = 1000,
  relative = 5,
  max_peaks = 200,
  normalized = T
)
}
\description{
Function used by library_generator to detect MS2 scans
}
