% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readMGF2.R
\name{readMGF2}
\alias{readMGF2}
\title{Extracting metadata and a list of spectra from mgf file}
\usage{
readMGF2(con)
}
\arguments{
\item{con}{Name of the input library, the file extension must be mgf}
}
\value{
\itemize{
  \item{sp:}{ List of all extracted spectra. Each spectrum is a data matrix with two columns: m/z and intensity}
  \item{metadata:}{ Data frame containing metadata of extracted scans. PEPMASS and RT are updated based on actually-detected scans. Following five columns are added: FILENAME, MSLEVEL, TIC, MASS_DEV, SCANNUMBER and SCANS}
}
}
\description{
The function extracts a dataframe of metadata and a list of spectra from mgf file or
}
