% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readMSP2.R
\name{readMSP2}
\alias{readMSP2}
\title{Extracting metadata and a list of spectra from msp file}
\usage{
readMSP2(con)
}
\arguments{
\item{con}{Name of the input library, the file extension must be msp}
}
\value{
\itemize{
  \item{sp:}{ List of all extracted spectra. Each spectrum is a data matrix with two columns: m/z and intensity}
  \item{metadata:}{ Data frame containing metadata of extracted scans.}
}
}
\description{
The function extracts a dataframe of metadata and a list of spectra from msp file
}
