% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_SmartMS2.R
\name{separated_peaks2}
\alias{separated_peaks2}
\title{Find indexes from "ranges": separated isomer peaks according to rt and intensity}
\usage{
separated_peaks2(ranges, rts, tics, rt_gap)
}
\description{
Find indexes from "ranges": separated isomer peaks according to rt and intensity
}
