% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_consensus.R
\name{updateSpecOneID}
\alias{updateSpecOneID}
\title{Genereate consensus library for a given compound ID}
\usage{
updateSpecOneID(
  listSpec,
  method = c("most_recent", "consensus", "consensus2", "common_peaks"),
  scans,
  consensus_window
)
}
\arguments{
\item{listSpec}{list of spectra}

\item{scans}{the "SCANS" from the metadata}
}
\description{
Genereate consensus library for a given compound ID
}
