% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeMGF2.R
\name{writeMGF2}
\alias{writeMGF2}
\title{Writing metadata and a list of spectra to mgf files}
\usage{
writeMGF2(library, con)
}
\arguments{
\item{library}{List that contains 2 elements.
\itemize{
  \item{sp:}{ List of all extracted spectra. Each spectrum is a data matrix with two columns: m/z and intensity}
  \item{metadata:}{ Data frame containing metadata of extracted scans. PEPMASS and RT are updated based on actually-detected scans. Following five columns are added: FILENAME, MSLEVEL, TIC, MASS_DEV, SCANNUMBER and SCANS}
}}

\item{con}{Name of the output library, the file extension must be mgf}
}
\description{
The function writes a dataframe of metadata and a list of spectra to a mgf file
}
\examples{

data(DRUG_THERMO_LIBRARY)

# Add new metadata "RESOLUTION = HIGH" to all scans:
library2$metadata$RESOLUTION = "HIGH"
# Write into a new mgf file:
writeMGF2(library2,"library_V2_bis.mgf")

}
