% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeMSP2.R
\name{writeMSP2}
\alias{writeMSP2}
\title{Writing metadata and a list of spectra to mgf files}
\usage{
writeMSP2(library, con)
}
\arguments{
\item{library}{List that contains 2 elements.
\itemize{
  \item{sp:}{ List of all extracted spectra. Each spectrum is a data matrix with two columns: m/z and intensity}
  \item{metadata:}{ Data frame containing metadata of extracted scans. PEPMASS and RT are updated based on actually-detected scans. Following five columns are added: FILENAME, MSLEVEL, TIC, MASS_DEV, SCANNUMBER and SCANS}
}}

\item{con}{Name of the output library, the file extension must be msp}
}
\description{
The function writes a dataframe of metadata and a list of spectra to a msp file
}
\examples{

data(DRUG_THERMO_LIBRARY)

# Add new metadata "RESOLUTION = HIGH" to all scans:
library2$metadata$RESOLUTION = "HIGH"
# Write into a new mgf file:
writeMSP(library2,"library.msp")

}
