% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_spectrum.r
\name{average_spectrum}
\alias{average_spectrum}
\title{Generating consensus spectra}
\usage{
average_spectrum(splist, ppm_window = 10, clean = F)
}
\arguments{
\item{splist}{List of extracted spectra. Each spectrum is a data matrix with two columns: m/z and intensity}

\item{ppm_window}{m/z tolerance window (in ppm) for spectra alignment}

\item{clean}{Logical. TRUE if peaks not present in all spectra are filtered.}
}
\value{
new_spectrum ~ The aligned, (filtered) and averaged consensus spectra: data matrix with two columns (m/z and intensity)
}
\description{
Internal function that merges a list of normalized spectra that belong to the same ID
}
