% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library_generator.r
\name{library_generator}
\alias{library_generator}
\title{Generating spectral library from raw LC-MS/MS chromatograms}
\usage{
library_generator(
  input_library = NULL,
  raw_data_files = NULL,
  metadata_file = NULL,
  polarity = c("Positive", "Negative"),
  mslevel = c(1, 2),
  add.adduct = TRUE,
  processing.algorithm = c("Default", "compMS2Miner", "RMassBank"),
  params.search = list(mz_search = 0.005, ppm_search = 10, rt_seach = 15, rt_gap = 30),
  params.ms.preprocessing = list(normalized = T, baseline = 1000, relative = 0.01,
    max_peaks = 200, recalibration = F),
  params.user = ""
)
}
\arguments{
\item{input_library}{Character or library object. If character, name of the library into which new scans are added, the file extension must be mgf; please set to empty string "" or NULL if the new library has no dependency with previous ones.}

\item{raw_data_files}{A character vector of LC-MS/MS file names from which scans are extracted. All files must have be in centroid-mode with mzML or mzMXL extension!}

\item{metadata_file}{A single character. It should be the metadata file name. The file should be txt, csv or xlsx format. The txt or csv file can be tab, comma or semi-colon separated. The xlsx file must be a single sheet. For all algorithms, the metadata must contain the column "ID" - a unique structure identifier. The column PEPMASS (targeted precursor mass) must be provided for Default and compMS2Miner. The column RT (targeted retention time in min) must be provided for compMS2Miner and optional for MergeION and RMassBank. Please include the column SMILES (structure identifier) for RMassBank algorithm. If RMassBank is used, the column FILENAME (chromatogram file with mzML or mzXML extension) must be provided for each compound telling the algorithm from which file compound can be found. Column FILENAME is optional for Default and compMS2Miner. Column ADDUCT is optional for all algorithms, if not provided, all input will be considered as M+H or M-H depending on polarity. Please specify the adduct type if metadata contains both positive and negative ions.}

\item{polarity}{A single character. Either "Positive" or "Negative". Ion mode of LC-MS/MS files.}

\item{mslevel}{A numeric vector. Must contain 2 (if only MS2 scans are extracted) and can be c(1,2) if isotopic pattern in MS1 scans are also extracted. Note: High-quality isotopic patterns in MS1 scans are useful for determining precursor formula!}

\item{add.adduct}{Logical. If TRUE, additional adduct types will be calculated based on precursor masses of "M+H" and "M-H" adducts in the input metadata: "M+2H", "M+Na","M+K","M+NH4","M+" will be searched for positive ion mode, "M-2H", "M+COO-", "M+Cl" and "M-" for negative ion mode. If FALSE, no additional adduct types will be searched.}

\item{processing.algorithm}{A single character. "Default", "compMS2Miner" or "RMassBank".}

\item{params.search}{Parameters for searching and collecting ions from chromatogram files in a list. These parameters define the tolerance window when input metadata is searched. The list must contain following elements:
\itemize{
 \item{mz_search:}{ Numeric. Absolute mass tolerance in Da.}
 \item{ppm_search:}{ Numeric. Absolute mass tolerance in ppm.} 
 \item{rt_search":}{ Numeric. Absolute retention time tolerance in second.}
 \item{rt_gap:}{ Numeric. Retention time gap in second - when two scans both match with an input structure, they are both recorded as isomeric features of the same identifier if they are separated by a certain retention time gap. Please set it to 10000 if no isomeric feature is picked. This parameter is not used for RMassBank.}
}}

\item{params.ms.preprocessing}{Paremters for post-processing scans found in chromatogram files in a list. It must contain:
\itemize{
 \item{normalized:}{ Logical. TRUE if the intensities of extracted spectra need to normalized so that the intensity of highest peak will be 100.}
 \item{baseline:}{ Numeric. Absolute intensity threshold that is considered as a mass peak and written into the library.}
 \item{relative:}{ Numeric between 0 and 100. Relative intensity threshold of the highest peak in each spectrum, peaks above both absolute and relative thresholds are saved in the library.}
 \item{max_peaks:}{ Integer higher than 3. Maximum number of peaks kept per spectrum from the highest peak.}
 \item{recalibration:}{ Logical. TRUE if recalibration is performed using RMassBank algorithm.}
}}

\item{params.user}{A single character. User name who processed the data.}
}
\value{
\itemize{
  \item{complete:}{ Entire spectra library (historical + newly added records) is a list object of two elements: "library$sp" ~ List of all extracted spectra. Each spectrum is a data matrix with two columns: m/z and intensity; "library$metadata" ~ Data frame containing metadata of extracted scans. PEPMASS and RT are updated based on scans detected in the chromatogram files. Following metadata columns are updated/added: FILENAME (which raw data file the scan is isolated), MSLEVEL (1 or 2), TIC, PEPMASS_DEV (ppm error for detected precursor mass) and SCANNUMBER (scan number in raw chromatogram). The last three columns were PARAM_SUBMIT_USER (user name), PARAM_ALGORITHM (algorithm of processing), PARAM_CREATION_TIME (date and time when the MS record was added) and SCANS (unique identifier for each record)}
  \item{current:}{ Temporary spectra library that only contains newly added scans.}
}
}
\description{
The function proposes three data processing algorithms to pick up MS1/MS2 scans from DDA or targeted mode LC-MS/MS data and merge them into a spectral library (new or existing).
}
\examples{


input_library = NULL # There's no historical spectral library. We create a brand new spectral library here,
raw_data_files <- list.files(system.file("spectra", package = "MergeION"),".mzML", full.names = TRUE)
metadata_file <- list.files(system.file(package = "MergeION"),".xlsx", full.names = TRUE)

polarity = "Positive"
mslevel= 2 # Only MS2 scans are extracted!
add.adduct = F # No additional adducts are searched besides M+H 

params.search = list(mz_search = 0.005, ppm_search = 10, rt_seach = 15, rt_gap = 30)
params.ms.preprocessing = list(normalized = T, baseline = 1000, relative =0.01, max_peaks = 200, recalibration = F)
params.user = "DANIEL"

processing.algorithm = "Default"
lib = library_generator(input_library, raw_data_files, metadata_file, 
                       polarity = "Positive", mslevel, add.adduct, processing.algorithm,
                       params.search, params.ms.preprocessing, params.user)
lib1 = lib$complete
save(lib1, file = "test_Default.RData") # Save the library as RData

processing.algorithm = "compMS2Miner"
lib = library_generator(input_library, raw_data_files, metadata_file, 
                       polarity = "Positive", mslevel, add.adduct, processing.algorithm,
                       params.search, params.ms.preprocessing, params.user)
lib2 = lib$complete
save(lib2, file = "test_compMS2Miner.RData")  # Save the library as RData

# Processing with RMassBank without recalibration:
processing.algorithm = "RMassBank"
lib = library_generator(input_library, raw_data_files, metadata_file, 
                      polarity = "Positive", mslevel, add.adduct, processing.algorithm,
                      params.search, params.ms.preprocessing, params.user)
lib3 = lib$complete
save(lib3, file = "test_RMassBank.RData")  # Save the library as RData

# Processing with RMassBank with the recalibration function:
params.ms.preprocessing$recalibration = TRUE
lib = library_generator(input_library, raw_data_files, metadata_file, 
                     polarity = "Positive", mslevel, add.adduct, processing.algorithm,
                     params.search, params.ms.preprocessing, params.user)
save(lib4, file = "test_RMassBank_calibrated.RData")  

}
\author{
Youzhong Liu, \email{YLiu186@ITS.JNJ.com}
}
