% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library_messar.R
\name{library_messar}
\alias{library_messar}
\title{Substructure recommendation for a "library" of unknown compounds}
\usage{
library_messar(
  library,
  type = c("drug", "metabolite"),
  tops = 5,
  max_peaks = 200,
  min_relative = 0.1,
  output = "library_messar_drug.mgf"
)
}
\arguments{
\item{library}{A list generated by the function library_generator() or the name of mgf spectral library file}

\item{type}{Character. Two algorithms focused either on "drug" or "metabolite"}

\item{tops}{Integer. Top recommended substructures in the output mgf file}

\item{max_peaks}{Integer. Top most intense peaks kept in the spectra}

\item{min_relative}{Numeric between 0.1 and 1. Minimum relative intensity of peaks to not be considered as noise.}

\item{output}{String. File name of output library.}
}
\value{
\itemize{
   \item{library_annotated:}{ Library annotated}
   \item{sub_extracted:}{ Library object that contain found scans.}
   \item{MGF FILE:}{ A MGF file with annotated substructures will be written in user's folder.}
}
}
\description{
The function recommends substructures for unknown spectra stored in a mgf file.
}
\examples{

data(DRUG_THERMO_LIBRARY)
results = library_messar(library2, type = "drug", tops = 5, output = "thermo_substructure.mgf")

}
\author{
Youzhong Liu, \email{Youzhong.Liu@uantwerpen.be}
}
