% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library_messar.R
\name{library_messar}
\alias{library_messar}
\title{Substructure recommendation for unknown spectrum}
\usage{
library_messar(
  query_spectrum = NULL,
  params.search = list(mz_search = 0.01, tops = 5),
  params.query.sp = list(prec_mz = 100, use_prec = T, compound.type = "Metabolite")
)
}
\arguments{
\item{query_spectrum}{Two-column data matrix. Two columns represent m/z and intensity of query tandem spectrum. At least 3 valid peaks should be provided}

\item{params.search}{General parameters for matching fragments to 
\itemize{
 \item{mz_search:}{ Numeric. Absolute mass tolerance in Da for fragment match.}
 \item{tops:}{ Integer. Top n substructures recommended.}
}}

\item{params.query.sp}{Parameters for matching an unknown spectra
\itemize{
  \item{prec_mz:}{ Numeric. Precursor mass of query spectrum. please put the correct value of precursor mass.}
  \item{use_prec:}{ Boolean. If set to TRUE, precursor mass is used to calculate the neutral loss.}
  \item{compound.type:}{ Character. Either "Metabolite" or "Drug". Different models are applied to Metabolite or Drug.}
}}
}
\value{
Recommended substructures (ordered) as a two-column matrix: SMILES code and score
}
\description{
The function recommends substructures for an unknown spectrum
}
\examples{

url_test = "https://raw.githubusercontent.com/daniellyz/MergeION2/master/inst/spectra/caffeine.txt"

test = read.table(url_test, header = F, sep=" ")

results = library_messar(query_spectrum = test, params.query.sp = list(prec_mz = 195.088, use_prec = T, compound.type = "Drug"))

}
\author{
Youzhong Liu, \email{Youzhong.Liu@uantwerpen.be}
}
