% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library_query.R
\name{library_query}
\alias{library_query}
\title{Spectral library query}
\usage{
library_query(
  input_library = NULL,
  query_ids = NULL,
  query_expression = "",
  query_spectrum = NULL,
  query_library = NULL,
  params.search = list(mz_search = 0.01, ppm_search = 10, rt_search = 15, rt_gap = 30),
  params.query.sp = list(prec_mz = 10000, use_prec = T, polarity = "Positive", method =
    "Cosine", min_frag_match = 6, min_score = 0, reaction_type = "Metabolic")
)
}
\arguments{
\item{input_library}{Character or a list object. If character, file name with extension mgf, msp or RData. If list, must contain "complete", "consensus" and "network".}

\item{query_ids}{Vector. Vectors of molecular ids that will be extracted from input library}

\item{query_expression}{Vector of characters or "". Vector of conditions used for querying the library. e.g. c("IONMODE=Positive","PEPMASS=325.19"). The left-hand side must match with the metadata items of the searched library.}

\item{query_spectrum}{Two-column data matrix. Two columns represent m/z and intensity of query tandem spectrum. At least 3 valid peaks should be provided.}

\item{query_library}{Character or a list object. Used when query_spectrum = NULL. If character, file name with extension mgf, msp, Rdata, cdf, mz(x)ml. If list, must contain metadata and sp, or library_generator output ("complete", "consensus" and "network")}

\item{params.search}{General parameters for searching spectral library
\itemize{
 \item{mz_search:}{ Numeric. Absolute mass tolerance in Da for fragment match.}
 \item{ppm_search:}{ Numeric. Mass tolerance in ppm for precursor mass match.} 
 \item{rt_search:}{ Numeric. Absolute retention time tolerance in second.}
 \item{rt_gap:}{ Numeric. Only used when querying a raw LC-MS/MS file. Retention time gap in second - when two scans both match with an input structure, they are both recorded as isomeric features of the same identifier if they are separated by a certain retention time gap. Please set it to 10000 if no isomeric feature is picked}
}}

\item{params.query.sp}{Parameters for matching an unknown spectra
\itemize{
  \item{prec_mz:}{ Numeric. Precursor mass of query spectrum.}
  \item{use_prec:}{ Boolean. If set to TRUE, precursor mass is used to "pre-query" the library.}
  \item{polarity:}{ Character. Ion mode of the query spectrum, either positive or negative.}
  \item{method:}{ Character. Similarity metrics.}
   \enumerate{
       \item{Precision:}{ Percentage of fragment or neutral loss matched in query spectrum.}
       \item{Recall:}{ Percentage of fragment or neutral loss matched in reference spectrum.}
       \item{F1:}{ Harmonic mean of precision and recall.}
       \item{Cosine:}{ COsine similarity score based on intensity vectors of fragments.}
       \item{Spearman:}{ Spearman similarity based on intensity ranks of fragments.}
       \item{MassBank:}{Similarity score used by MassBank using weighted pearson score}
       \item{NIST:}{Similarity score used by NIST using weighted pearson score}}
   \item{min_frag_match}{ Minimum matched peaks (or corresponding neutral losses) to allow a match.}
   \item{min_score}{ Minimum similarity score to be considered as a successful query.}
   \item{reaction_type:}{ Character. Either "Metabolic" and "Chemical". Type of transformation list used to annotate mass difference between connected features in molecular network.}
}}
}
\value{
\itemize{
  \item{complete:}{ Complete spectra library filtered by both query experession and query spectrum}
  \item{consensus:}{ Consensus spectra library filtered by both query experession and query spectrum.}
  \item{network:}{ Molecular network filtered by both query experession and query spectrum. The query spectrum is also connected to the molecular network}
}
}
\description{
The function queries the input (reference) spectral library based on a query expression and/or (a) query MS/MS spectra or query library or LC-MS/MS file
}
\examples{

}
\author{
Youzhong Liu, \email{Youzhong.Liu@uantwerpen.be}
}
