% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library_reader.R
\name{library_reader}
\alias{library_reader}
\title{Read and extract an existing library}
\usage{
library_reader(raw_library)
}
\arguments{
\item{raw_library}{Library file name (character) or object generated by library_generator. If character, the file extension must be mgf, msp or rdata}
}
\value{
\itemize{
  \item{sp:}{ List of all extracted spectra. Each spectrum is a data matrix with two columns: m/z and intensity} 
  \item{metadata:}{ Data frame containing metadata of extracted scans.}
}
}
\description{
The function creates or extracts a list object containing dataframe of metadata and a list of spectra from library object
}
