% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library_visualizer.R
\name{library_visualizer}
\alias{library_visualizer}
\title{Visualize selected mass spectra in the spectral library}
\usage{
library_visualizer(
  input_library,
  id = input_library$metadata$ID[1],
  query_spectrum = NULL
)
}
\arguments{
\item{input_library}{A list generated by the function library_generator() or the name of mgf/msp/RData spectral library file}

\item{id}{Compound id of the spectra to be plotted (both most recent MS1 and MS2 spectra will be plotted!)}

\item{query_spectrum}{Two-column data matrix. Optional for mirror plot. Two columns represent m/z and intensity of query tandem spectrum.}
}
\description{
The function plots mass spectra of selected compound in a library
}
\examples{

data(DRUG_THERMO_LIBRARY)

# Search library using query command lines:
query = library_manager(library2,query=c("IONMODE=Positive","RT=1.2"), logical="AND", rt_search=6)

# Visualize the first compound found:
library_visualizer(query$SELECTED, id = query$ID_SELECTED[1])

# Plot against query spectrum:
query_spectrum = cbind(c(71.084, 89.6, 455.178), c(90,40,39))
library_visualizer(library2, id = "Selexipag", query_spectrum) 

}
