% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library_visualizer.R
\name{library_visualizer}
\alias{library_visualizer}
\title{Visualize selected mass spectra in the spectral library}
\usage{
library_visualizer(library, id = library$metadata$ID[1], png.out = F,
  show.legend = T)
}
\arguments{
\item{library}{A list generated by the function library_generator() or library_similarity()$SELECTED}

\item{id}{Compound id of the spectra to be plotted (both most recent MS1 and MS2 spectra will be plotted!)}

\item{png.out}{Boolean. True if plotted spectra are exported as png images!}
}
\description{
The function plots mass spectra of selected compound in a library
}
\examples{

data(DRUG_THERMO_LIBRARY)

# Search library using query command lines:
query = library_manager(library2,query=c("IONMODE=Positive","RT=1.2"), logical="AND", rt_search=6)

# Visualize the first compound found:
library_visualizer(query$SELECTED, id = query$ID_SELECTED[1])

}
