% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_visualizer.R
\name{library_visualizer_interactive}
\alias{library_visualizer_interactive}
\title{Interactive spectral plot}
\usage{
library_visualizer_interactive(
  plotData,
  x = "m/z",
  y = Intensity,
  colorVar = ID,
  facetVar = MSLEVEL
)
}
\arguments{
\item{plotData}{A data frame used to plot}

\item{x}{x-axis variable quote or unquoted}

\item{y}{y-axis variable}

\item{facetVar}{the variable used to facet the graph by row}

\item{coloVar}{color variable}
}
\description{
Interactive spectral plot
}
\author{
Yingjie Zhang
}
