% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_generator.R
\name{network_generator}
\alias{network_generator}
\title{Create feature-based molecular networks}
\usage{
network_generator(
  raw_data_file,
  input_library = NULL,
  input_matrix = NULL,
  polarity = c("Positive", "Negative"),
  params.search = list(mz_search = 0.005, ppm_search = 10, rt_seach = 15),
  params.ms.preprocessing = list(baseline = 1000, relative = 0.01, max_peaks = 200),
  params.similarity = list(method = "Cosine", use.prec = FALSE, min_frag_match = 6)
)
}
\arguments{
\item{raw_data_file}{Character. The LC-MS/MS file in mzML/mzXML format or MS/MS spectral summary file in mgf format. Check following link for more information about MS/MS spectral summary file: https://ccms-ucsd.github.io/GNPSDocumentation/featurebasedmolecularnetworking-with-mzmine2/}

\item{input_library}{A list generated by the function library_generator() or the name of mgf/msp/RData spectral library file.}

\item{input_matrix}{Transformed spectral library by the function library2matrix. Used for MassBank or GNPS spectra library search. Please set to NULL otherwise}

\item{polarity}{character. Either "Positive" or "Negative". Ion mode of the LC-MS/MS file.}

\item{params.search}{Parameters for clustering ions from raw_data_file and for searching in the spectral library. The list must contain following elements:
\itemize{
 \item{mz_search:}{ Numeric. Absolute mass tolerance in Da.}
 \item{ppm_search:}{ Numeric. Absolute mass tolerance in ppm.} 
 \item{rt_search":}{ Numeric. Absolute retention time tolerance in second.}
}}

\item{params.ms.preprocessing}{Parameters for post-processing scans in raw_data_file:
\itemize{
 \item{baseline:}{ Numeric. Absolute intensity threshold that is considered as a mass peak and written into the library.}
 \item{relative:}{ Numeric between 0 and 100. Relative intensity threshold of the highest peak in each spectrum, peaks above both absolute and relative thresholds are saved in the library.}
 \item{max_peaks:}{ Integer higher than 3. Maximum number of peaks kept per spectrum from the highest peak.}
}}

\item{params.similarity}{Parameters for MS/MS spectral similarity determination
\itemize{
 \item{method:}{Characeter.Similarity metrics for networking and spectral library search. Must be "Matches", "Cosine", "Spearman", "Euclidean", "MassBank", "NIST". Please check function library_similarity for more details.}
 \item{use.prec:}{Boolean. Used for spectral library search. If set to TRUE, precursor mass is used to "pre-query" the library}
 \item{min_frag_match:}{Integer. Minimum matched peaks (or corresponding neutral losses) to make a match.}
}}
}
\description{
Generate feature-based molecular networks from raw LC-MS/MS file or spectral summary file, automatically annotate features in a spectral library
}
