% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_consensus.R
\name{process_consensus}
\alias{process_consensus}
\title{Extracting or generating representative or consensus scans for each compound ID}
\usage{
process_consensus(
  input_library,
  method = c("most_recent", "consensus", "consensus2", "common_peaks")[1],
  consensus_window = 0.01,
  relative = 0.01,
  max_peaks = 200
)
}
\description{
Function used by library_generator to generate consensus scans
}
