% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_metadata.R
\name{process_metadata}
\alias{process_metadata}
\title{Create additional metadata for library generation}
\usage{
process_metadata(
  ref,
  processing.algorithm = c("Default", "compMS2Miner", "RMassBank"),
  polarity = c("Positive", "Negative"),
  add.adduct = T
)
}
\description{
Function used by library_generator() to improve and adapt input metadata
}
