% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_query.R
\name{process_query}
\alias{process_query}
\title{Searching and managing the spectral library}
\usage{
process_query(
  library0,
  query = "",
  ppm_search = 20,
  rt_search = 12,
  sepString = "@xxx@"
)
}
\arguments{
\item{query}{query expression - two ways to specify drop}

\item{setString}{string used to seperate paramter name and expression
\itemize{
\item{equal:}{use e.g. 'IONMODE=Positive' to do exact search}
\item{regex:}{use e.g. 'IONMODE@xxx@^Pos' to search IONMODE begins with 'Pos'}
}}
}
\description{
The function is used by library_query() to select scans based on a query expression.
}
