% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_similarity.R
\name{process_similarity}
\alias{process_similarity}
\title{Searching a query spectrum in a spectral library}
\usage{
process_similarity(
  query_spectrum,
  polarity = "Positive",
  prec_mz = 100,
  use.prec = FALSE,
  input_library = NULL,
  method = c("Matches", "Dot", "Cosine", "Spearman", "MassBank", "NIST"),
  prec_ppm_search = 10,
  frag_mz_search = 0.005,
  min_frag_match = 6
)
}
\description{
The function calculates spectral similarity of a query spectrum to an existing spectral library
}
\author{
Youzhong Liu, \email{Youzhong.Liu@uantwerpen.be}
}
